requirements.txt
numpy==1.23.5
Pillow==10.1.0
PyYAML==6.0.1
scikit_learn==1.2.0
scipy==1.11.4
torch==1.13.0+cu116
torchvision==0.14.0+cu116
